requirements.txt
# HTTP 클라이언트 (비동기)
httpx==0.27.0

# 환경변수
python-dotenv==1.0.1

# 데이터 처리 (conda로 설치 권장: conda install pandas numpy)
pandas>=2.0.0
numpy>=1.24.0

# 데이터베이스
# SQLite는 Python 내장 (별도 설치 불필요)

# 시각화
plotly==5.22.0

# 서빙 UI
streamlit==1.36.0

# AI 코치
anthropic==0.31.0

# 유틸리티
tqdm==4.66.4          # 진행률 표시
loguru==0.7.2         # 로깅
aiofiles==23.2.1      # 비동기 파일 I/O

# 테스트
pytest==8.2.2
pytest-asyncio==0.23.7